refactor: extract CommunityRefreshStage from IncrementalIndexer
